backend/requirements.txt
langchain
langchain-core==0.3.74
langchain-community==0.3.0
langchain-openai==0.3.30
langchain-google-genai==2.0.0
langgraph==0.6.0
python-dotenv